+++ mkutils.lua
@@ -338,7 +338,10 @@ env.Make:add("test","test the variables:  ${tex4ht_sty_par} ${htlatex} ${input} 
 local htlatex = require "make4ht-htlatex"
 env.Make:add("htlatex", htlatex.htlatex
 ,{correct_exit=0})
-
+env.Make:add("htttex", htlatex.httex, {
+  htlatex = "etex",
+  correct_exit=0
+})
 
 env.Make:add("latexmk", function(par)
   local settings = get_filter_settings "htlatex" or {}